Rename: architecture -> back 변경
